Split up the repositories in repositories and services where the services do all the things that do not change stuff.
