+++ src/BoardgameController.php
@@ -5,6 +5,7 @@ use \DateTimeImmutable;
 class BoardgameController
 {
     public BoardgameRepository $repository;
+    public BoardgameService $service;
 
     /**
      * Initialises the boardgame repository
@@ -12,6 +13,7 @@ class BoardgameController
     public function __construct()
     {
         $this->repository = new BoardgameRepository();
+        $this->service = new BoardgameService();
     }
 
     /**
@@ -20,7 +22,7 @@ class BoardgameController
      */
     public function boardgameIndex()
     {
-        $boardgames = $this->repository->getAllBoardgames();
+        $boardgames = $this->service->getAllBoardgames();
         include_once 'html/boardgameindex.html';
     }
 
@@ -31,7 +33,7 @@ class BoardgameController
      */
     public function showBoardgame(int $id)
     {
-        $boardgame = $this->repository->getBoardgameById($id);
+        $boardgame = $this->service->getBoardgameById($id);
         include_once 'html/Boardgame.html';
     }
