+++ src/ItemService.php
@@ -5,23 +5,23 @@ use \DateTimeImmutable;
 class ItemService
 {
 
-    public BookRepository $bookRepository;
-    public MagazineRepository $magazineRepository;
-    public BoardgameRepository $boardgameRepository;
+    public BookService $bookService;
+    public MagazineService $magazineService;
+    public BoardgameService $boardgameService;
 
     public function __construct()
     {
-        $this->bookRepository = new BookRepository();
-        $this->magazineRepository = new MagazineRepository();
-        $this->boardgameRepository = new BoardgameRepository();
+        $this->bookService = new BookService();
+        $this->magazineService = new MagazineService();
+        $this->boardgameService = new BoardgameService();
     }
 
     public function getAllItems()
     {
         $items = [];
-        $books = $this->bookRepository->getAll();
-        $magazines = $this->magazineRepository->getAllMagazines();
-        $boardgames = $this->boardgameRepository->getAllBoardgames();
+        $books = $this->bookService->getAll();
+        $magazines = $this->magazineService->getAllMagazines();
+        $boardgames = $this->boardgameService->getAllBoardgames();
         $items = array_merge($items, $books, $magazines, $boardgames);
         return $items;
     }
@@ -30,12 +30,18 @@ class ItemService
     {
         $search = $data['item'];
         $items = [];
-        $books = $this->bookRepository->searchBooks($search);
-        $magazines = $this->magazineRepository->searchMagazines($search);
-        $boardgames = $this->boardgameRepository->searchBoardgames($search);
+        $books = $this->bookService->searchBooks($search);
+        $magazines = $this->magazineService->searchMagazines($search);
+        $boardgames = $this->boardgameService->searchBoardgames($search);
 
         $items = array_merge($items, $books, $magazines, $boardgames);
         return $items;
+
+    }
+
+    //Meak how?
+    public function returnItem()
+    {
 
     }
