+++ src/MagazineController.php
@@ -7,12 +7,15 @@ class MagazineController
 
     public MagazineRepository $repository;
 
+    public MagazineService $service;
+
     /**
      * Initialises the magazine repository
      */
     public function __construct()
     {
         $this->repository = new MagazineRepository();
+        $this->service = new MagazineService();
     }
 
     /**
@@ -21,7 +24,7 @@ class MagazineController
      */
     public function magazineIndex()
     {
-        $magazines = $this->repository->getAllMagazines();
+        $magazines = $this->service->getAllMagazines();
         include_once 'html/magazineindex.html';
     }
 
@@ -32,7 +35,7 @@ class MagazineController
      */
     public function showMagazine(int $id)
     {
-        $magazine = $this->repository->returnMagazineById($id);
+        $magazine = $this->service->returnMagazineById($id);
         include_once 'html/Magazine.html';
     }
