+++ src/MagazineRepository.php
@@ -24,26 +24,26 @@ class MagazineRepository
         $this->queryBuilder->insert($keyValuePairs);
     }
 
-    /**
-     * Retrieves all magazines from the database
-     * @return array|null
-     */
-    public function getAllMagazines()
-    {
-        return $magazines = $this->queryBuilder->select(['*'])->get();
+    // /**
+    //  * Retrieves all magazines from the database
+    //  * @return array|null
+    //  */
+    // public function getAllMagazines()
+    // {
+    //     return $magazines = $this->queryBuilder->select(['*'])->get();
 
-    }
+    // }
 
-    /**
-     * Gets one magazine from the database with the given id
-     * @param int $id
-     */
-    public function returnMagazineById(int $id)
-    {
-        $magazine = $this->queryBuilder->select(['*'])->where(['Id' => $id])->get();
+    // /**
+    //  * Gets one magazine from the database with the given id
+    //  * @param int $id
+    //  */
+    // public function returnMagazineById(int $id)
+    // {
+    //     $magazine = $this->queryBuilder->select(['*'])->where(['Id' => $id])->get();
 
-        return $magazine[0];
-    }
+    //     return $magazine[0];
+    // }
 
     /**
      * Removes a magazine from the database with the given id
@@ -55,18 +55,18 @@ class MagazineRepository
         $this->queryBuilder->remove($id);
     }
 
-    /**
-     * Searches the magazines database table on title publisher and editor and returns an array of results
-     * @param string $search
-     * @return array
-     */
-    public function searchMagazines(string $search)
-    {
-        $magazines = [];
-        $titles = $this->queryBuilder->select(['*'])->where(['Title' => $search])->get();
-        $publishers = $this->queryBuilder->select(['*'])->where(['Publisher' => $search])->get();
-        $editors = $this->queryBuilder->select(['*'])->where(['Publisher' => $search])->get();
-        $magazines = array_merge($magazines, $titles, $publishers, $editors);
-        return $magazines;
-    }
+    // /**
+    //  * Searches the magazines database table on title publisher and editor and returns an array of results
+    //  * @param string $search
+    //  * @return array
+    //  */
+    // public function searchMagazines(string $search)
+    // {
+    //     $magazines = [];
+    //     $titles = $this->queryBuilder->select(['*'])->where(['Title' => $search])->get();
+    //     $publishers = $this->queryBuilder->select(['*'])->where(['Publisher' => $search])->get();
+    //     $editors = $this->queryBuilder->select(['*'])->where(['Publisher' => $search])->get();
+    //     $magazines = array_merge($magazines, $titles, $publishers, $editors);
+    //     return $magazines;
+    // }
 }
